+++ app/src/main/java/com/example/tms/EightHW/CollectionPhones.kt
@@ -4,5 +4,4 @@ import com.example.tms.EightHW.Database.entity.Phone
 
 object CollectionPhones {
     var phones = listOf<Phone>()
-    var wasUpdated = true
 }
