+++ app/src/main/res/layout/fragment_add_phone.xml
@@ -54,6 +54,14 @@
             />
 
     </com.google.android.material.textfield.TextInputLayout>
+    <Button
+        android:id="@+id/button_add_phone"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="@string/add_phone"
+        android:layout_marginTop="400dp"
+        />
+
 
 
 </LinearLayout>
